deepfake_module/backend1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
tensorflow==2.18.0
numpy
pillow
